requirements.txt
langchain
langchain-core
langchain-groq
requests
python-dotenv
beautifulsoup4
playwright